+++ cmd/neofs-node/accounting.go
@@ -3,7 +3,6 @@ package main
 import (
 	accountingGRPC "github.com/nspcc-dev/neofs-api-go/v2/accounting/grpc"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client/balance"
-	accountingTransportGRPC "github.com/nspcc-dev/neofs-node/pkg/network/transport/accounting/grpc"
 	accountingService "github.com/nspcc-dev/neofs-node/pkg/services/accounting"
 	accounting "github.com/nspcc-dev/neofs-node/pkg/services/accounting/morph"
 )
@@ -16,7 +15,7 @@ func initAccountingService(c *cfg) {
 	balanceMorphWrapper, err := balance.NewFromMorph(c.cfgMorph.client, c.shared.basics.balanceSH, 0)
 	fatalOnErr(err)
 
-	server := accountingTransportGRPC.New(
+	server := accountingService.New(
 		accountingService.NewSignService(
 			&c.key.PrivateKey,
 			accountingService.NewResponseService(
