--- pkg/network/transport/object/grpc/replication.go
@@ -1,283 +1,0 @@
-package object
-
-import (
-	"bytes"
-	"context"
-	"encoding/binary"
-	"errors"
-	"fmt"
-
-	objectv2 "github.com/nspcc-dev/neofs-api-go/v2/object"
-	objectGRPC "github.com/nspcc-dev/neofs-api-go/v2/object/grpc"
-	refsv2 "github.com/nspcc-dev/neofs-api-go/v2/refs"
-	refs "github.com/nspcc-dev/neofs-api-go/v2/refs/grpc"
-	status "github.com/nspcc-dev/neofs-api-go/v2/status/grpc"
-	objectcore "github.com/nspcc-dev/neofs-node/pkg/core/object"
-	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
-	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
-	neofscrypto "github.com/nspcc-dev/neofs-sdk-go/crypto"
-	neofsecdsa "github.com/nspcc-dev/neofs-sdk-go/crypto/ecdsa"
-	"github.com/nspcc-dev/neofs-sdk-go/object"
-	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
-)
-
-// Replicate serves neo.fs.v2.object.ObjectService/Replicate RPC.
-func (s *Server) Replicate(_ context.Context, req *objectGRPC.ReplicateRequest) (*objectGRPC.ReplicateResponse, error) {
-	if req.Object == nil {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code: codeInternal, Message: "binary object field is missing/empty",
-		}}, nil
-	}
-
-	if req.Object.ObjectId == nil || len(req.Object.ObjectId.Value) == 0 {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code: codeInternal, Message: "ID field is missing/empty in the object field",
-		}}, nil
-	}
-
-	if req.Signature == nil {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code: codeInternal, Message: "missing object signature field",
-		}}, nil
-	}
-
-	if len(req.Signature.Key) == 0 {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code: codeInternal, Message: "public key field is missing/empty in the object signature field",
-		}}, nil
-	}
-
-	if len(req.Signature.Sign) == 0 {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code: codeInternal, Message: "signature value is missing/empty in the object signature field",
-		}}, nil
-	}
-
-	switch scheme := req.Signature.Scheme; scheme {
-	default:
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code:    codeInternal,
-			Message: "unsupported scheme in the object signature field",
-		}}, nil
-	case
-		refs.SignatureScheme_ECDSA_SHA512,
-		refs.SignatureScheme_ECDSA_RFC6979_SHA256,
-		refs.SignatureScheme_ECDSA_RFC6979_SHA256_WALLET_CONNECT:
-	}
-
-	hdr := req.Object.GetHeader()
-	if hdr == nil {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code:    codeInternal,
-			Message: "missing header field in the object field",
-		}}, nil
-	}
-
-	gCnrMsg := hdr.GetContainerId()
-	if gCnrMsg == nil {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code:    codeInternal,
-			Message: "missing container ID field in the object header field",
-		}}, nil
-	}
-
-	var cnr cid.ID
-	var cnrMsg refsv2.ContainerID
-	err := cnrMsg.FromGRPCMessage(gCnrMsg)
-	if err == nil {
-		err = cnr.ReadFromV2(cnrMsg)
-	}
-	if err != nil {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code:    codeInternal,
-			Message: fmt.Sprintf("invalid container ID in the object header field: %v", err),
-		}}, nil
-	}
-
-	var pubKey neofscrypto.PublicKey
-	switch req.Signature.Scheme {
-	// other cases already checked above
-	case refs.SignatureScheme_ECDSA_SHA512:
-		pubKey = new(neofsecdsa.PublicKey)
-		err = pubKey.Decode(req.Signature.Key)
-		if err != nil {
-			return &objectGRPC.ReplicateResponse{Status: &status.Status{
-				Code:    codeInternal,
-				Message: "invalid ECDSA public key in the object signature field",
-			}}, nil
-		}
-	case refs.SignatureScheme_ECDSA_RFC6979_SHA256:
-		pubKey = new(neofsecdsa.PublicKeyRFC6979)
-		err = pubKey.Decode(req.Signature.Key)
-		if err != nil {
-			return &objectGRPC.ReplicateResponse{Status: &status.Status{
-				Code:    codeInternal,
-				Message: "invalid ECDSA public key in the object signature field",
-			}}, nil
-		}
-	case refs.SignatureScheme_ECDSA_RFC6979_SHA256_WALLET_CONNECT:
-		pubKey = new(neofsecdsa.PublicKeyWalletConnect)
-		err = pubKey.Decode(req.Signature.Key)
-		if err != nil {
-			return &objectGRPC.ReplicateResponse{Status: &status.Status{
-				Code:    codeInternal,
-				Message: "invalid ECDSA public key in the object signature field",
-			}}, nil
-		}
-	}
-	if !pubKey.Verify(req.Object.ObjectId.Value, req.Signature.Sign) {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code:    codeInternal,
-			Message: "signature mismatch in the object signature field",
-		}}, nil
-	}
-
-	var clientInCnr, serverInCnr bool
-	err = s.node.ForEachContainerNodePublicKeyInLastTwoEpochs(cnr, func(pubKey []byte) bool {
-		if !serverInCnr {
-			serverInCnr = s.node.IsOwnPublicKey(pubKey)
-		}
-		if !clientInCnr {
-			clientInCnr = bytes.Equal(pubKey, req.Signature.Key)
-		}
-		return !clientInCnr || !serverInCnr
-	})
-	if err != nil {
-		if errors.Is(err, apistatus.ErrContainerNotFound) {
-			return &objectGRPC.ReplicateResponse{Status: &status.Status{
-				Code:    codeContainerNotFound,
-				Message: "failed to check server's compliance to object's storage policy: object's container not found",
-			}}, nil
-		}
-
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code:    codeInternal,
-			Message: fmt.Sprintf("failed to apply object's storage policy: %v", err),
-		}}, nil
-	} else if !serverInCnr {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code: codeAccessDenied, Message: "server does not match the object's storage policy",
-		}}, nil
-	} else if !clientInCnr {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code: codeAccessDenied, Message: "client does not match the object's storage policy",
-		}}, nil
-	}
-
-	// TODO(@cthulhu-rider): avoid decoding the object completely
-	obj, err := objectFromMessage(req.Object)
-	if err != nil {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code:    codeInternal,
-			Message: fmt.Sprintf("invalid object field: %v", err),
-		}}, nil
-	}
-
-	err = s.node.VerifyAndStoreObject(*obj)
-	if err != nil {
-		return &objectGRPC.ReplicateResponse{Status: &status.Status{
-			Code:    codeInternal,
-			Message: fmt.Sprintf("failed to verify and store object locally: %v", err),
-		}}, nil
-	}
-
-	resp := new(objectGRPC.ReplicateResponse)
-	if req.GetSignObject() {
-		resp.ObjectSignature, err = s.metaInfoSignature(*obj)
-		if err != nil {
-			return &objectGRPC.ReplicateResponse{Status: &status.Status{
-				Code:    codeInternal,
-				Message: fmt.Sprintf("failed to sign object meta information: %v", err),
-			}}, nil
-		}
-	}
-
-	return resp, nil
-}
-
-func objectFromMessage(gMsg *objectGRPC.Object) (*object.Object, error) {
-	var msg objectv2.Object
-	err := msg.FromGRPCMessage(gMsg)
-	if err != nil {
-		return nil, err
-	}
-
-	return object.NewFromV2(&msg), nil
-}
-
-func (s *Server) metaInfoSignature(o object.Object) ([]byte, error) {
-	firstObj := o.GetFirstID()
-	if o.HasParent() && firstObj.IsZero() {
-		// object itself is the first one
-		firstObj = o.GetID()
-	}
-	prevObj := o.GetPreviousID()
-
-	var deleted []oid.ID
-	var locked []oid.ID
-	typ := o.Type()
-	switch typ {
-	case object.TypeTombstone:
-		var t object.Tombstone
-		err := t.Unmarshal(o.Payload())
-		if err != nil {
-			return nil, fmt.Errorf("reading tombstoned objects: %w", err)
-		}
-
-		deleted = t.Members()
-	case object.TypeLock:
-		var l object.Lock
-		err := l.Unmarshal(o.Payload())
-		if err != nil {
-			return nil, fmt.Errorf("reading locked objects: %w", err)
-		}
-
-		locked = make([]oid.ID, l.NumberOfMembers())
-		l.ReadMembers(locked)
-	default:
-	}
-
-	currentBlock := s.nmState.CurrentBlock()
-	currentEpochDuration := s.nmState.CurrentEpochDuration()
-	firstBlock := (uint64(currentBlock)/currentEpochDuration + 1) * currentEpochDuration
-	secondBlock := firstBlock + currentEpochDuration
-	thirdBlock := secondBlock + currentEpochDuration
-
-	firstMeta := objectcore.EncodeReplicationMetaInfo(o.GetContainerID(), o.GetID(), firstObj, prevObj, o.PayloadSize(), typ, deleted, locked, firstBlock, s.mNumber)
-	secondMeta := objectcore.EncodeReplicationMetaInfo(o.GetContainerID(), o.GetID(), firstObj, prevObj, o.PayloadSize(), typ, deleted, locked, secondBlock, s.mNumber)
-	thirdMeta := objectcore.EncodeReplicationMetaInfo(o.GetContainerID(), o.GetID(), firstObj, prevObj, o.PayloadSize(), typ, deleted, locked, thirdBlock, s.mNumber)
-
-	var firstSig neofscrypto.Signature
-	var secondSig neofscrypto.Signature
-	var thirdSig neofscrypto.Signature
-
-	err := firstSig.Calculate(s.signer, firstMeta)
-	if err != nil {
-		return nil, fmt.Errorf("signature failure: %w", err)
-	}
-	err = secondSig.Calculate(s.signer, secondMeta)
-	if err != nil {
-		return nil, fmt.Errorf("signature failure: %w", err)
-	}
-	err = thirdSig.Calculate(s.signer, thirdMeta)
-	if err != nil {
-		return nil, fmt.Errorf("signature failure: %w", err)
-	}
-
-	firstSigV2 := new(refsv2.Signature)
-	firstSig.WriteToV2(firstSigV2)
-	secondSigV2 := new(refsv2.Signature)
-	secondSig.WriteToV2(secondSigV2)
-	thirdSigV2 := new(refsv2.Signature)
-	thirdSig.WriteToV2(thirdSigV2)
-
-	res := make([]byte, 0, 4+firstSigV2.StableSize()+4+secondSigV2.StableSize()+4+thirdSigV2.StableSize())
-	res = binary.LittleEndian.AppendUint32(res, uint32(firstSigV2.StableSize()))
-	res = append(res, firstSigV2.StableMarshal(nil)...)
-	res = binary.LittleEndian.AppendUint32(res, uint32(secondSigV2.StableSize()))
-	res = append(res, secondSigV2.StableMarshal(nil)...)
-	res = binary.LittleEndian.AppendUint32(res, uint32(thirdSigV2.StableSize()))
-	res = append(res, thirdSigV2.StableMarshal(nil)...)
-
-	return res, nil
-}
