+++ pkg/services/object/server.go
@@ -1,44 +1,535 @@
 package object
 
 import (
+	"bytes"
 	"context"
+	"encoding/binary"
+	"errors"
+	"fmt"
+	"io"
 
-	"github.com/nspcc-dev/neofs-api-go/v2/object"
+	v2object "github.com/nspcc-dev/neofs-api-go/v2/object"
+	protoobject "github.com/nspcc-dev/neofs-api-go/v2/object/grpc"
+	refsv2 "github.com/nspcc-dev/neofs-api-go/v2/refs"
+	refs "github.com/nspcc-dev/neofs-api-go/v2/refs/grpc"
+	protostatus "github.com/nspcc-dev/neofs-api-go/v2/status/grpc"
+	"github.com/nspcc-dev/neofs-node/pkg/core/netmap"
+	objectcore "github.com/nspcc-dev/neofs-node/pkg/core/object"
 	"github.com/nspcc-dev/neofs-node/pkg/services/util"
+	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
+	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
+	neofscrypto "github.com/nspcc-dev/neofs-sdk-go/crypto"
+	neofsecdsa "github.com/nspcc-dev/neofs-sdk-go/crypto/ecdsa"
+	"github.com/nspcc-dev/neofs-sdk-go/object"
+	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
 )
 
 // GetObjectStream is an interface of NeoFS API v2 compatible object streamer.
 type GetObjectStream interface {
 	util.ServerStream
-	Send(*object.GetResponse) error
+	Send(*v2object.GetResponse) error
 }
 
 // GetObjectRangeStream is an interface of NeoFS API v2 compatible payload range streamer.
 type GetObjectRangeStream interface {
 	util.ServerStream
-	Send(*object.GetRangeResponse) error
+	Send(*v2object.GetRangeResponse) error
 }
 
 // SearchStream is an interface of NeoFS API v2 compatible search streamer.
 type SearchStream interface {
 	util.ServerStream
-	Send(*object.SearchResponse) error
+	Send(*v2object.SearchResponse) error
 }
 
 // PutObjectStream is an interface of NeoFS API v2 compatible client's object streamer.
 type PutObjectStream interface {
-	Send(*object.PutRequest) error
-	CloseAndRecv() (*object.PutResponse, error)
+	Send(*v2object.PutRequest) error
+	CloseAndRecv() (*v2object.PutResponse, error)
 }
 
 // ServiceServer is an interface of utility
 // serving v2 Object service.
 type ServiceServer interface {
-	Get(*object.GetRequest, GetObjectStream) error
+	Get(*v2object.GetRequest, GetObjectStream) error
 	Put(context.Context) (PutObjectStream, error)
-	Head(context.Context, *object.HeadRequest) (*object.HeadResponse, error)
-	Search(*object.SearchRequest, SearchStream) error
-	Delete(context.Context, *object.DeleteRequest) (*object.DeleteResponse, error)
-	GetRange(*object.GetRangeRequest, GetObjectRangeStream) error
-	GetRangeHash(context.Context, *object.GetRangeHashRequest) (*object.GetRangeHashResponse, error)
+	Head(context.Context, *v2object.HeadRequest) (*v2object.HeadResponse, error)
+	Search(*v2object.SearchRequest, SearchStream) error
+	Delete(context.Context, *v2object.DeleteRequest) (*v2object.DeleteResponse, error)
+	GetRange(*v2object.GetRangeRequest, GetObjectRangeStream) error
+	GetRangeHash(context.Context, *v2object.GetRangeHashRequest) (*v2object.GetRangeHashResponse, error)
+}
+
+// Various NeoFS protocol status codes.
+const (
+	codeInternal          = uint32(1024*protostatus.Section_SECTION_FAILURE_COMMON) + uint32(protostatus.CommonFail_INTERNAL)
+	codeAccessDenied      = uint32(1024*protostatus.Section_SECTION_OBJECT) + uint32(protostatus.Object_ACCESS_DENIED)
+	codeContainerNotFound = uint32(1024*protostatus.Section_SECTION_CONTAINER) + uint32(protostatus.Container_CONTAINER_NOT_FOUND)
+)
+
+// Node represents NeoFS storage node that is served by [Server].
+type Node interface {
+	// ForEachContainerNodePublicKeyInLastTwoEpochs iterates over all nodes matching
+	// the referenced container's storage policy at the current and the previous
+	// NeoFS epochs, and passes their public keys into f. IterateContainerNodeKeys
+	// breaks without an error when f returns false. Keys may be repeated.
+	//
+	// Returns [apistatus.ErrContainerNotFound] if referenced container was not
+	// found.
+	ForEachContainerNodePublicKeyInLastTwoEpochs(cid.ID, func(pubKey []byte) bool) error
+
+	// IsOwnPublicKey checks whether given pubKey assigned to Node in the NeoFS
+	// network map.
+	IsOwnPublicKey(pubKey []byte) bool
+
+	// VerifyAndStoreObject checks whether given object has correct format and, if
+	// so, saves it into local object storage of the Node. StoreObject is called
+	// only when the Node complies with the container's storage policy.
+	VerifyAndStoreObject(object.Object) error
+}
+
+type server struct {
+	srv ServiceServer
+
+	node    Node
+	signer  neofscrypto.Signer
+	mNumber uint32
+	nmState netmap.StateDetailed
+}
+
+// New provides protoobject.ObjectServiceServer for the given parameters.
+func New(c ServiceServer, magicNumber uint32, node Node, signer neofscrypto.Signer, nmState netmap.StateDetailed) protoobject.ObjectServiceServer {
+	return &server{
+		srv:     c,
+		node:    node,
+		signer:  signer,
+		mNumber: magicNumber,
+		nmState: nmState,
+	}
+}
+
+// Put opens internal Object service Put stream and overtakes data from gRPC stream to it.
+func (s *server) Put(gStream protoobject.ObjectService_PutServer) error {
+	stream, err := s.srv.Put(gStream.Context())
+	if err != nil {
+		return err
+	}
+
+	for {
+		req, err := gStream.Recv()
+		if err != nil {
+			if errors.Is(err, io.EOF) {
+				resp, err := stream.CloseAndRecv()
+				if err != nil {
+					return err
+				}
+
+				return gStream.SendAndClose(resp.ToGRPCMessage().(*protoobject.PutResponse))
+			}
+
+			return err
+		}
+
+		putReq := new(v2object.PutRequest)
+		if err := putReq.FromGRPCMessage(req); err != nil {
+			return err
+		}
+
+		if err := stream.Send(putReq); err != nil {
+			if errors.Is(err, util.ErrAbortStream) {
+				resp, err := stream.CloseAndRecv()
+				if err != nil {
+					return err
+				}
+
+				return gStream.SendAndClose(resp.ToGRPCMessage().(*protoobject.PutResponse))
+			}
+
+			return err
+		}
+	}
+}
+
+// Delete converts gRPC DeleteRequest message and passes it to internal Object service.
+func (s *server) Delete(ctx context.Context, req *protoobject.DeleteRequest) (*protoobject.DeleteResponse, error) {
+	delReq := new(v2object.DeleteRequest)
+	if err := delReq.FromGRPCMessage(req); err != nil {
+		return nil, err
+	}
+
+	resp, err := s.srv.Delete(ctx, delReq)
+	if err != nil {
+		return nil, err
+	}
+
+	return resp.ToGRPCMessage().(*protoobject.DeleteResponse), nil
+}
+
+// Head converts gRPC HeadRequest message and passes it to internal Object service.
+func (s *server) Head(ctx context.Context, req *protoobject.HeadRequest) (*protoobject.HeadResponse, error) {
+	searchReq := new(v2object.HeadRequest)
+	if err := searchReq.FromGRPCMessage(req); err != nil {
+		return nil, err
+	}
+
+	resp, err := s.srv.Head(ctx, searchReq)
+	if err != nil {
+		return nil, err
+	}
+
+	return resp.ToGRPCMessage().(*protoobject.HeadResponse), nil
+}
+
+// GetRangeHash converts gRPC GetRangeHashRequest message and passes it to internal Object service.
+func (s *server) GetRangeHash(ctx context.Context, req *protoobject.GetRangeHashRequest) (*protoobject.GetRangeHashResponse, error) {
+	hashRngReq := new(v2object.GetRangeHashRequest)
+	if err := hashRngReq.FromGRPCMessage(req); err != nil {
+		return nil, err
+	}
+
+	resp, err := s.srv.GetRangeHash(ctx, hashRngReq)
+	if err != nil {
+		return nil, err
+	}
+
+	return resp.ToGRPCMessage().(*protoobject.GetRangeHashResponse), nil
+}
+
+type getStreamerV2 struct {
+	protoobject.ObjectService_GetServer
+}
+
+func (s *getStreamerV2) Send(resp *v2object.GetResponse) error {
+	return s.ObjectService_GetServer.Send(
+		resp.ToGRPCMessage().(*protoobject.GetResponse),
+	)
+}
+
+// Get converts gRPC GetRequest message and server-side stream and overtakes its data
+// to gRPC stream.
+func (s *server) Get(req *protoobject.GetRequest, gStream protoobject.ObjectService_GetServer) error {
+	getReq := new(v2object.GetRequest)
+	if err := getReq.FromGRPCMessage(req); err != nil {
+		return err
+	}
+
+	return s.srv.Get(
+		getReq,
+		&getStreamerV2{
+			ObjectService_GetServer: gStream,
+		},
+	)
+}
+
+type getRangeStreamerV2 struct {
+	protoobject.ObjectService_GetRangeServer
+}
+
+func (s *getRangeStreamerV2) Send(resp *v2object.GetRangeResponse) error {
+	return s.ObjectService_GetRangeServer.Send(
+		resp.ToGRPCMessage().(*protoobject.GetRangeResponse),
+	)
+}
+
+// GetRange converts gRPC GetRangeRequest message and server-side stream and overtakes its data
+// to gRPC stream.
+func (s *server) GetRange(req *protoobject.GetRangeRequest, gStream protoobject.ObjectService_GetRangeServer) error {
+	getRngReq := new(v2object.GetRangeRequest)
+	if err := getRngReq.FromGRPCMessage(req); err != nil {
+		return err
+	}
+
+	return s.srv.GetRange(
+		getRngReq,
+		&getRangeStreamerV2{
+			ObjectService_GetRangeServer: gStream,
+		},
+	)
+}
+
+type searchStreamerV2 struct {
+	protoobject.ObjectService_SearchServer
+}
+
+func (s *searchStreamerV2) Send(resp *v2object.SearchResponse) error {
+	return s.ObjectService_SearchServer.Send(
+		resp.ToGRPCMessage().(*protoobject.SearchResponse),
+	)
+}
+
+// Search converts gRPC SearchRequest message and server-side stream and overtakes its data
+// to gRPC stream.
+func (s *server) Search(req *protoobject.SearchRequest, gStream protoobject.ObjectService_SearchServer) error {
+	searchReq := new(v2object.SearchRequest)
+	if err := searchReq.FromGRPCMessage(req); err != nil {
+		return err
+	}
+
+	return s.srv.Search(
+		searchReq,
+		&searchStreamerV2{
+			ObjectService_SearchServer: gStream,
+		},
+	)
+}
+
+// Replicate serves neo.fs.v2.object.ObjectService/Replicate RPC.
+func (s *server) Replicate(_ context.Context, req *protoobject.ReplicateRequest) (*protoobject.ReplicateResponse, error) {
+	if req.Object == nil {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code: codeInternal, Message: "binary object field is missing/empty",
+		}}, nil
+	}
+
+	if req.Object.ObjectId == nil || len(req.Object.ObjectId.Value) == 0 {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code: codeInternal, Message: "ID field is missing/empty in the object field",
+		}}, nil
+	}
+
+	if req.Signature == nil {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code: codeInternal, Message: "missing object signature field",
+		}}, nil
+	}
+
+	if len(req.Signature.Key) == 0 {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code: codeInternal, Message: "public key field is missing/empty in the object signature field",
+		}}, nil
+	}
+
+	if len(req.Signature.Sign) == 0 {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code: codeInternal, Message: "signature value is missing/empty in the object signature field",
+		}}, nil
+	}
+
+	switch scheme := req.Signature.Scheme; scheme {
+	default:
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code:    codeInternal,
+			Message: "unsupported scheme in the object signature field",
+		}}, nil
+	case
+		refs.SignatureScheme_ECDSA_SHA512,
+		refs.SignatureScheme_ECDSA_RFC6979_SHA256,
+		refs.SignatureScheme_ECDSA_RFC6979_SHA256_WALLET_CONNECT:
+	}
+
+	hdr := req.Object.GetHeader()
+	if hdr == nil {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code:    codeInternal,
+			Message: "missing header field in the object field",
+		}}, nil
+	}
+
+	gCnrMsg := hdr.GetContainerId()
+	if gCnrMsg == nil {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code:    codeInternal,
+			Message: "missing container ID field in the object header field",
+		}}, nil
+	}
+
+	var cnr cid.ID
+	var cnrMsg refsv2.ContainerID
+	err := cnrMsg.FromGRPCMessage(gCnrMsg)
+	if err == nil {
+		err = cnr.ReadFromV2(cnrMsg)
+	}
+	if err != nil {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code:    codeInternal,
+			Message: fmt.Sprintf("invalid container ID in the object header field: %v", err),
+		}}, nil
+	}
+
+	var pubKey neofscrypto.PublicKey
+	switch req.Signature.Scheme {
+	// other cases already checked above
+	case refs.SignatureScheme_ECDSA_SHA512:
+		pubKey = new(neofsecdsa.PublicKey)
+		err = pubKey.Decode(req.Signature.Key)
+		if err != nil {
+			return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+				Code:    codeInternal,
+				Message: "invalid ECDSA public key in the object signature field",
+			}}, nil
+		}
+	case refs.SignatureScheme_ECDSA_RFC6979_SHA256:
+		pubKey = new(neofsecdsa.PublicKeyRFC6979)
+		err = pubKey.Decode(req.Signature.Key)
+		if err != nil {
+			return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+				Code:    codeInternal,
+				Message: "invalid ECDSA public key in the object signature field",
+			}}, nil
+		}
+	case refs.SignatureScheme_ECDSA_RFC6979_SHA256_WALLET_CONNECT:
+		pubKey = new(neofsecdsa.PublicKeyWalletConnect)
+		err = pubKey.Decode(req.Signature.Key)
+		if err != nil {
+			return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+				Code:    codeInternal,
+				Message: "invalid ECDSA public key in the object signature field",
+			}}, nil
+		}
+	}
+	if !pubKey.Verify(req.Object.ObjectId.Value, req.Signature.Sign) {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code:    codeInternal,
+			Message: "signature mismatch in the object signature field",
+		}}, nil
+	}
+
+	var clientInCnr, serverInCnr bool
+	err = s.node.ForEachContainerNodePublicKeyInLastTwoEpochs(cnr, func(pubKey []byte) bool {
+		if !serverInCnr {
+			serverInCnr = s.node.IsOwnPublicKey(pubKey)
+		}
+		if !clientInCnr {
+			clientInCnr = bytes.Equal(pubKey, req.Signature.Key)
+		}
+		return !clientInCnr || !serverInCnr
+	})
+	if err != nil {
+		if errors.Is(err, apistatus.ErrContainerNotFound) {
+			return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+				Code:    codeContainerNotFound,
+				Message: "failed to check server's compliance to object's storage policy: object's container not found",
+			}}, nil
+		}
+
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code:    codeInternal,
+			Message: fmt.Sprintf("failed to apply object's storage policy: %v", err),
+		}}, nil
+	} else if !serverInCnr {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code: codeAccessDenied, Message: "server does not match the object's storage policy",
+		}}, nil
+	} else if !clientInCnr {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code: codeAccessDenied, Message: "client does not match the object's storage policy",
+		}}, nil
+	}
+
+	// TODO(@cthulhu-rider): avoid decoding the object completely
+	obj, err := objectFromMessage(req.Object)
+	if err != nil {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code:    codeInternal,
+			Message: fmt.Sprintf("invalid object field: %v", err),
+		}}, nil
+	}
+
+	err = s.node.VerifyAndStoreObject(*obj)
+	if err != nil {
+		return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+			Code:    codeInternal,
+			Message: fmt.Sprintf("failed to verify and store object locally: %v", err),
+		}}, nil
+	}
+
+	resp := new(protoobject.ReplicateResponse)
+	if req.GetSignObject() {
+		resp.ObjectSignature, err = s.metaInfoSignature(*obj)
+		if err != nil {
+			return &protoobject.ReplicateResponse{Status: &protostatus.Status{
+				Code:    codeInternal,
+				Message: fmt.Sprintf("failed to sign object meta information: %v", err),
+			}}, nil
+		}
+	}
+
+	return resp, nil
+}
+
+func objectFromMessage(gMsg *protoobject.Object) (*object.Object, error) {
+	var msg v2object.Object
+	err := msg.FromGRPCMessage(gMsg)
+	if err != nil {
+		return nil, err
+	}
+
+	return object.NewFromV2(&msg), nil
+}
+
+func (s *server) metaInfoSignature(o object.Object) ([]byte, error) {
+	firstObj := o.GetFirstID()
+	if o.HasParent() && firstObj.IsZero() {
+		// object itself is the first one
+		firstObj = o.GetID()
+	}
+	prevObj := o.GetPreviousID()
+
+	var deleted []oid.ID
+	var locked []oid.ID
+	typ := o.Type()
+	switch typ {
+	case object.TypeTombstone:
+		var t object.Tombstone
+		err := t.Unmarshal(o.Payload())
+		if err != nil {
+			return nil, fmt.Errorf("reading tombstoned objects: %w", err)
+		}
+
+		deleted = t.Members()
+	case object.TypeLock:
+		var l object.Lock
+		err := l.Unmarshal(o.Payload())
+		if err != nil {
+			return nil, fmt.Errorf("reading locked objects: %w", err)
+		}
+
+		locked = make([]oid.ID, l.NumberOfMembers())
+		l.ReadMembers(locked)
+	default:
+	}
+
+	currentBlock := s.nmState.CurrentBlock()
+	currentEpochDuration := s.nmState.CurrentEpochDuration()
+	firstBlock := (uint64(currentBlock)/currentEpochDuration + 1) * currentEpochDuration
+	secondBlock := firstBlock + currentEpochDuration
+	thirdBlock := secondBlock + currentEpochDuration
+
+	firstMeta := objectcore.EncodeReplicationMetaInfo(o.GetContainerID(), o.GetID(), firstObj, prevObj, o.PayloadSize(), typ, deleted, locked, firstBlock, s.mNumber)
+	secondMeta := objectcore.EncodeReplicationMetaInfo(o.GetContainerID(), o.GetID(), firstObj, prevObj, o.PayloadSize(), typ, deleted, locked, secondBlock, s.mNumber)
+	thirdMeta := objectcore.EncodeReplicationMetaInfo(o.GetContainerID(), o.GetID(), firstObj, prevObj, o.PayloadSize(), typ, deleted, locked, thirdBlock, s.mNumber)
+
+	var firstSig neofscrypto.Signature
+	var secondSig neofscrypto.Signature
+	var thirdSig neofscrypto.Signature
+
+	err := firstSig.Calculate(s.signer, firstMeta)
+	if err != nil {
+		return nil, fmt.Errorf("signature failure: %w", err)
+	}
+	err = secondSig.Calculate(s.signer, secondMeta)
+	if err != nil {
+		return nil, fmt.Errorf("signature failure: %w", err)
+	}
+	err = thirdSig.Calculate(s.signer, thirdMeta)
+	if err != nil {
+		return nil, fmt.Errorf("signature failure: %w", err)
+	}
+
+	firstSigV2 := new(refsv2.Signature)
+	firstSig.WriteToV2(firstSigV2)
+	secondSigV2 := new(refsv2.Signature)
+	secondSig.WriteToV2(secondSigV2)
+	thirdSigV2 := new(refsv2.Signature)
+	thirdSig.WriteToV2(thirdSigV2)
+
+	res := make([]byte, 0, 4+firstSigV2.StableSize()+4+secondSigV2.StableSize()+4+thirdSigV2.StableSize())
+	res = binary.LittleEndian.AppendUint32(res, uint32(firstSigV2.StableSize()))
+	res = append(res, firstSigV2.StableMarshal(nil)...)
+	res = binary.LittleEndian.AppendUint32(res, uint32(secondSigV2.StableSize()))
+	res = append(res, secondSigV2.StableMarshal(nil)...)
+	res = binary.LittleEndian.AppendUint32(res, uint32(thirdSigV2.StableSize()))
+	res = append(res, thirdSigV2.StableMarshal(nil)...)
+
+	return res, nil
 }
